Pythonic update
