+++ fitsnap3/__main__.py
@@ -28,9 +28,12 @@
 #     Laura Swiler (Sandia National Labs)
 # <!-----------------END-HEADER------------------------------------->
 
-from .fitsnap import FitSnap
-from .parallel_tools import pt
-from .io.output import output
+from fitsnap3.fitsnap import FitSnap
+from fitsnap3.parallel_tools import ParallelTools
+from fitsnap3.io.output import output
+
+
+pt = ParallelTools()
 
 
 @pt.single_timeit
