+++ fitsnap3/calculators/lammps_snap.py
@@ -1,8 +1,12 @@
 import ctypes
-from ..calculators.calculator import Calculator
-from ..parallel_tools import pt
-from ..io.input import config
+from fitsnap3.calculators.calculator import Calculator
+from fitsnap3.parallel_tools import ParallelTools
+from fitsnap3.io.input import Config
 import numpy as np
+
+
+config = Config()
+pt = ParallelTools()
 
 
 class LammpsSnap(Calculator):
