+++ fitsnap3/io/output.py
@@ -1,5 +1,9 @@
-from ..io.outputs.output_factory import output_factory
-from ..io.input import config
+from fitsnap3.io.outputs.output_factory import output_factory
+from fitsnap3.io.input import Config
+
+
+config = Config()
+
 
 if __name__.split(".")[-1] == "output":
     output = output_factory(config.sections["OUTFILE"].output_style)
