+++ fitsnap3/io/outputs/outputs.py
@@ -1,9 +1,13 @@
-from ..input import config
-from ...parallel_tools import pt
+from fitsnap3.io.input import Config
+from fitsnap3.parallel_tools import ParallelTools
 from contextlib import contextmanager
 import gzip
 from datetime import datetime
 import logging
+
+
+config = Config()
+pt = ParallelTools()
 
 
 class Output:
