+++ fitsnap3/io/outputs/pace.py
@@ -1,8 +1,12 @@
-from .outputs import Output, optional_open
-from ...parallel_tools import pt
+from fitsnap3.io.outputs.outputs import Output, optional_open
+from fitsnap3.parallel_tools import ParallelTools
 from datetime import datetime
-from ..input import config
+from fitsnap3.io.input import Config
 import numpy as np
+
+
+config = Config()
+pt = ParallelTools()
 
 
 class Pace(Output):
