+++ fitsnap3/io/outputs/template_output.py
@@ -1,4 +1,4 @@
-from .outputs import Output
+from fitsnap3.io.outputs.outputs import Output
 """Methods you may or must override in new output"""
