+++ fitsnap3/io/sections/calculator_sections/basic_calculator.py
@@ -1,5 +1,5 @@
 """Basic Class"""
-from ..sections import Section
+from fitsnap3.io.sections.sections import Section
 
 
 class Basic(Section):
