+++ fitsnap3/io/sections/extras.py
@@ -1,5 +1,8 @@
-from .sections import Section
-from ...parallel_tools import pt
+from fitsnap3.io.sections.sections import Section
+from fitsnap3.parallel_tools import ParallelTools
+
+
+pt = ParallelTools()
 
 
 class Extras(Section):
