+++ fitsnap3/io/sections/groups.py
@@ -1,8 +1,12 @@
 """Groups Class"""
-from .sections import Section, output
+from fitsnap3.io.sections.sections import Section
 from pandas import read_csv
 from os import path
-from ...parallel_tools import pt
+from fitsnap3.parallel_tools import ParallelTools
+
+
+pt = ParallelTools()
+
 
 def _str_2_fun(some_list):
     for i, item in enumerate(some_list):
