+++ fitsnap3/io/sections/outfile.py
@@ -1,6 +1,9 @@
 from os import path
-from .sections import Section
-from ...parallel_tools import pt
+from fitsnap3.io.sections.sections import Section
+from fitsnap3.parallel_tools import ParallelTools
+
+
+pt = ParallelTools()
 
 
 class Outfile(Section):
