+++ fitsnap3/io/sections/path.py
@@ -1,6 +1,9 @@
-from .sections import Section
+from fitsnap3.io.sections.sections import Section
 from os import path
-from ...parallel_tools import pt
+from fitsnap3.parallel_tools import ParallelTools
+
+
+pt = ParallelTools()
 
 
 class Path(Section):
