+++ fitsnap3/io/sections/section_factory.py
@@ -1,6 +1,8 @@
-from .sections import Section, pt
-from ..error import ExitFunc
-pt.get_subclasses(__name__, __file__, Section)
+from fitsnap3.io.sections.sections import Section
+from fitsnap3.io.error import ExitFunc
+from fitsnap3.parallel_tools import ParallelTools
+
+pt = ParallelTools()
 
 
 def new_section(section, config, args):
