+++ fitsnap3/io/sections/solver_sections/lasso.py
@@ -1,4 +1,4 @@
-from ..sections import Section
+from fitsnap3.io.sections.sections import Section
 
 
 class Lasso(Section):
