+++ fitsnap3/parallel_output.py
@@ -1,0 +1,19 @@
+from fitsnap3.parallel_tools import ParallelTools
+
+
+pt = ParallelTools()
+
+
+class Output:
+
+    def __init__(self):
+        self.none = None
+
+    @staticmethod
+    def screen(*args, **kw):
+        pt.single_print(*args, **kw)
+
+    @staticmethod
+    def exception(err):
+        # There is almost never a reason to use this
+        pt.exception(err)
