+++ fitsnap3/solvers/scalapack.py
@@ -1,10 +1,15 @@
-from .solver import Solver
-from ..parallel_tools import pt
-from ..io.input import config
+from fitsnap3.solvers.solver import Solver
+from fitsnap3.parallel_tools import ParallelTools
+from fitsnap3.io.input import Config
 import numpy as np
 
+
+config = Config()
+pt = ParallelTools()
+
+
 try:
-    from ..lib.scalapack_solver.scalapack import lstsq, dummy_lstsq
+    from fitsnap3.lib.scalapack_solver.scalapack import lstsq, dummy_lstsq
 
     class ScaLAPACK(Solver):
