+++ fitsnap3/solvers/svd.py
@@ -1,9 +1,13 @@
-from .solver import Solver
-from ..parallel_tools import pt
-from ..io.input import config
+from fitsnap3.solvers.solver import Solver
+from fitsnap3.parallel_tools import ParallelTools
+from fitsnap3.io.input import Config
 from scipy.linalg import lstsq
 import numpy as np
 from sys import float_info as fi
+
+
+config = Config()
+pt = ParallelTools()
 
 
 class SVD(Solver):
