+++ fitsnap3/solvers/template_solver.py
@@ -1,6 +1,9 @@
-from .solver import Solver
-from ..parallel_tools import pt
+from fitsnap3.solvers.solver import Solver
+from fitsnap3.parallel_tools import ParallelTools
 """Methods you may or must override in new solvers"""
+
+
+pt = ParallelTools()
 
 
 class Template(Solver):
