+++ fitsnaprun/__main__.py
@@ -1,0 +1,4 @@
+from fitsnap3.__main__ import main
+
+if __name__ == "__main__":
+    main()
